Fix: no outline found with PDFs
ttOutline() (in org.ebookdroid.droids.mupdf.codec.MuPdfOutline)
fails to find any outline if the argument `outline` is negative.

The JNI part of org.ebookdroid.droids.mupdf.codec.MuPdfOutline had
the following code:

    return (jlong) (long) doc->outline;

But this returns negative value if the most significant bit
of outline handle is 1, because `sizeof long` is 4 (32 bit)
and outline handler itself is a 32-bit pointer.